+++ redditrepostsleuth/core/celery/basetasks.py
@@ -49,6 +49,7 @@ class RedditTask(Task):
         self.reddit = RedditManager(get_reddit_instance(self.config))
         self.uowm = SqlAlchemyUnitOfWorkManager(get_db_engine(self.config))
         self.event_logger = EventLogging(config=self.config)
+        self.notification_svc = NotificationService(self.config)
         self.response_handler = ResponseHandler(self.reddit, self.uowm, self.event_logger, live_response=self.config.live_responses)
 
 class AdminTask(Task):